+++ src/main/java/janggi/board/Board.java
@@ -1,6 +1,8 @@
 package janggi.board;
 
 import janggi.GameState;
+import janggi.board.palace.Palace;
+import janggi.board.palace.PalaceGenerator;
 import janggi.piece.Piece;
 import janggi.piece.PieceType;
 import janggi.piece.Team;
