+++ src/main/java/janggi/board/palace/PalaceGenerator.java
@@ -1,4 +1,4 @@
-package janggi.board;
+package janggi.board.palace;
 
 import janggi.position.Position;
