+++ src/test/java/janggi/piece/PieceTest.java
@@ -2,7 +2,7 @@ package janggi.piece;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import janggi.board.Palace;
+import janggi.board.palace.Palace;
 import janggi.position.Position;
 import java.util.List;
 import java.util.Map;
