+++ src/test/java/janggi/piece/multiplemovepiece/CannonTest.java
@@ -5,8 +5,8 @@ import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import janggi.board.Board;
-import janggi.board.Palace;
-import janggi.board.PalaceGenerator;
+import janggi.board.palace.Palace;
+import janggi.board.palace.PalaceGenerator;
 import janggi.piece.Piece;
 import janggi.piece.PieceType;
 import janggi.piece.Team;
